requirements.txt
ephem
Pillow
numpy
astropy
git+https://github.com/esa/tetra3.git
